+++ src/Graph/Trees/BallTree.php
@@ -66,7 +66,7 @@ class BallTree implements BST, Spatial
                 . " to form a leaf node, $maxLeafSize given.");
         }
 
-        if (isset($kernel) and !in_array(DataType::CONTINUOUS, $kernel->compatibility())) {
+        if ($kernel and !in_array(DataType::CONTINUOUS, $kernel->compatibility())) {
             throw new InvalidArgumentException('Distance kernel must be'
                 . ' compatible with continuous features.');
         }
@@ -130,7 +130,7 @@ class BallTree implements BST, Spatial
             throw new InvalidArgumentException('Tree requires a labeled dataset.');
         }
 
-        if (!$dataset->homogeneous() or $dataset->columnType(0) !== DataType::CONTINUOUS) {
+        if ($dataset->columnType(0) !== DataType::CONTINUOUS or !$dataset->homogeneous()) {
             throw new InvalidArgumentException('This tree only works with'
                 . ' continuous features.');
         }